requirements.txt
backcall==0.1.0
Click==7.0
crab==0.5.0
cycler==0.10.0
decorator==4.4.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
joblib==0.13.2
jupyter-client==5.3.1
jupyter-core==4.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
pandas==0.23.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylzma==0.5.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
scikit-learn==0.21.3
scikit-surprise==1.0.6
scipy==1.3.1
six==1.12.0
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
